+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Avoid audio initialization warning
+- Do not auto-resume manually paused audio
 
 ## [0.4.0] - 2025-11-26